requirements.txt
django==4.2.2
python-dateutil==2.8.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
psycopg2-binary==2.9.6
pytest-django==4.4.0
pytest-factoryboy==2.1.0
mixer==7.2.1
model-bakery==1.3.2
